Update mainScript.py
Added the reward matrix.  It may not be the correct strategy, here are are few thoughts, they are also comments in the function I added.

1) The for loop is there for it so we can always change the values if need be.
2) I chose a "heavy" penalty term for having no connection because I am not sure if you can use null, I'm pretty sure you can't.
3) I chose a light penalty for moving from one node to another because if there was no penalty I would suspect a potential infinite loop.  The small penalty hopefully encourages exploration.

And ideally these changes would be set up that you would accept my changes.  It seems to do that I have to create a new branch but I don't think that is what I want to do.  So I'm guessing these changes will automatically go through.